moving thumbnail management code to gui
it will be moved from generic path based mechanism to id based
