+++ src/gui/desktop/utils/thumbnail_manager.cpp
