+++ src/gui/desktop/utils/thumbnails_cache.cpp
