+++ src/gui/gui_tests.cmake
@@ -13,6 +13,8 @@ addTestTarget(gui
                     desktop/models/flat_model.cpp
                     desktop/utils/model_index_utils.cpp
                     desktop/quick_views/selection_manager_component.cpp
+                    desktop/utils/thumbnail_manager.cpp
+                    desktop/utils/thumbnails_cache.cpp
 
                     # model tests:
                     unit_tests/model/aphoto_info_model_tests.cpp
@@ -28,6 +30,8 @@ addTestTarget(gui
                     # utils:
                     unit_tests/utils/model_index_utils_tests.cpp
                     unit_tests/utils/selection_manager_component_tests.cpp
+                    unit_tests/utils/thumbnails_manager_tests.cpp
+                    unit_tests/utils/thumbnails_cache_tests.cpp
 
                     # main()
                     unit_tests/main.cpp
