+++ src/unit_tests_utils/mock_thumbnails_generator.hpp
@@ -5,11 +5,11 @@
 
 #include <gmock/gmock.h>
 
-#include <core/ithumbnails_cache.hpp>
+#include <core/ithumbnails_generator.hpp>
 
 struct MockThumbnailsGenerator: IThumbnailsGenerator
 {
-    MOCK_METHOD(QImage, generate, (const QString &, const IThumbnailsCache::ThumbnailParameters& params), (override));
+    MOCK_METHOD(QImage, generate, (const QString &, const ThumbnailParameters& params), (override));
 };
 
 #endif
